# internal/api/openapi.yaml
openapi: 3.1.0
info:
  title: Tentacron
  version: "1"
  summary: Orchestration and resolvent API for renewable-energy modelling workflows.
  description: |
    Submit a model payload that still contains resolvent objects; tentacron
    resolves them through the configured resource APIs, forwards the completed
    payload to the target and, for asynchronous targets, polls the target's job
    to completion. Every request is durable and auditable.

    Authenticate every `/v1` call with the `X-API-Key` header. A key's role
    decides whether it reads only its own requests (`client`) or every request
    (`admin`). The API changes only additively within `/v1`; clients must ignore
    fields they do not know.
  license:
    name: MIT
    identifier: MIT
servers:
  - url: /
security:
  - ApiKeyAuth: []
...
paths:
  /v1/requests:
    post:
      tags: [requests]
      summary: Submit a request
      description: |
        Persists the request and answers `202` at once. Poll `GET /v1/requests/{id}`
        for progress. An `Idempotency-Key` (at most 255 bytes, scoped to the
        client) makes a resubmission of the identical request return the stored
        one instead of creating a duplicate; reusing the key with a different
        target or payload answers `409`.
      operationId: createRequest
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRequest"
      responses:
        "202":
          description: Accepted; the request's id and initial state.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Accepted"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          description: The Idempotency-Key was already used with a different target or payload.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "413":
          description: The body exceeds `server.max_body_bytes`.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "415":
          description: "`Content-Type: application/json` is required."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "422":
          description: The target is not configured.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    get:
      tags: [requests]
      summary: List requests
      description: |
        Newest first: the caller's own requests, or every client's for an admin
        key. A full page carries `next_cursor`; pass it back as `cursor` with the
        same filters for the next page. No total count is returned.
      operationId: listRequests
      parameters:
        - name: state
          in: query
          schema:
            $ref: "#/components/schemas/State"
        - name: target
          in: query
          schema:
            type: string
        - name: client
          in: query
          description: Admin keys only (a client may name itself).
          schema:
            type: string
        - name: since
          in: query
          description: Inclusive lower bound on `created_at`, RFC 3339.
          schema:
            type: string
            format: date-time
        - name: until
          in: query
          description: Exclusive upper bound on `created_at`, RFC 3339.
          schema:
            type: string
            format: date-time
        - name: limit
          in: query
          description: Page size, default 50, capped at 200.
          schema:
            type: integer
            minimum: 1
        - name: cursor
          in: query
          description: The `next_cursor` of the previous page. Opaque.
          schema:
            type: string
      responses:
        "200":
          description: One page of requests.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobList"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /v1/schedules:
    post:
      tags: [schedules]
      summary: Create a recurring submission
      description: |
        The same target and payload, submitted at every time the cron
        expression names (five fields, or `@hourly`, `@daily`, `@weekly`,
        `@monthly`, `@yearly`, `@every <duration>`), interpreted in
        `timezone` (an IANA name, default UTC). Runs are ordinary requests
        created with the idempotency key `schedule:<id>:<due time>`, default
        `options.cache: refresh`, and never earlier than their due time. The
        payload is inspected like a dry run, so a schedule whose every run
        would fail is refused.
      operationId: createSchedule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScheduleRequest"
      responses:
        "201":
          description: The schedule with its first due time.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Schedule"
        "400":
          description: Malformed body or payload, an invalid `cron`, `timezone` or `priority`, or a payload the dry run rejects (`invalid_payload`, `unknown_resolvent`, `target_error`).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          description: The target is not configured.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    get:
      tags: [schedules]
      summary: List schedules
      description: The caller's schedules, oldest first, at most 100. An admin key lists every client's, or one client's with `client`.
      operationId: listSchedules
      parameters:
        - name: client
          in: query
          schema: { type: string }
          description: Admin keys only.
      responses:
        "200":
          description: The schedules.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/Schedule" }
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/schedules/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string }
    get:
      tags: [schedules]
      summary: Read a schedule
      operationId: getSchedule
      responses:
        "200":
          description: The schedule.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Schedule"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [schedules]
      summary: Delete a schedule
      description: Runs already created stay as ordinary requests.
      operationId: deleteSchedule
      responses:
        "204":
          description: Deleted.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/schedules/{id}/runs:
    get:
      tags: [schedules]
      summary: List a schedule's runs
      description: The requests the schedule created, newest first, with the request list's `state`, `limit` and `cursor` parameters.
      operationId: listScheduleRuns
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
        - name: state
          in: query
          schema: { $ref: "#/components/schemas/State" }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
        - name: cursor
          in: query
          schema: { type: string }
      responses:
        "200":
          description: One page of runs.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobList"
        "400":
          description: Invalid `state`, `limit` or `cursor`.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  parameters:
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      schema:
        type: string
        maxLength: 255
  responses:
    BadRequest:
      description: Malformed body, missing field or invalid parameter.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Missing or unknown API key.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: No such request — or another client's request, which is indistinguishable.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    State:
      type: string
      enum: [received, resolving, forwarding, awaiting_target, completed, failed, cancelled]
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum: [invalid_json, missing_field, unauthorized, unknown_target, idempotency_conflict,
                     payload_too_large, unsupported_media_type, not_found, method_not_allowed,
                     not_cancellable, invalid_parameter, internal]
            message:
              type: string
    CreateRequest:
      type: object
      required: [target, payload]
      properties:
        target:
          type: string
          description: A target name from the configuration.
        payload:
          type: object
          description: The body to resolve and forward. Resolvent objects carry a `type` starting with `resolvent-`.
        priority:
          type: integer
          minimum: -10
          maximum: 10
          default: 0
          description: Higher priorities are claimed first; capped per key.
        options:
          type: object
          description: Per-request processing choices.
          properties:
            cache:
              type: string
              enum: [use, bypass, refresh]
              default: use
              description: "`bypass` neither reads nor writes the series cache; `refresh` fetches fresh and writes."
        not_before:
          type: string
          format: date-time
          description: Delays the run until this time (RFC 3339, at most 30 days ahead). The request is accepted at once and stays `received` until then; it can be cancelled meanwhile.
        api_key:
          type: string
          deprecated: true
          description: Fallback for the `X-API-Key` header; the header wins when both are present.
    Accepted:
      type: object
      required: [id, state, links]
      properties:
        id: { type: string }
        state: { $ref: "#/components/schemas/State" }
        links:
          type: object
          properties:
            self: { type: string }
    Job:
      type: object
      required: [id, target, state, attempts, created_at, updated_at, result, error]
      properties:
        id: { type: string }
        target: { type: string }
        state: { $ref: "#/components/schemas/State" }
        attempts: { type: integer }
        priority:
          type: integer
          description: Present when not the default 0.
        options:
          type: object
          description: Present when any option differs from its default.
          properties:
            cache: { type: string, enum: [bypass, refresh] }
        target_job_id:
          type: string
          description: The asynchronous target's own job id, once accepted.
        not_before:
          type: string
          format: date-time
          description: Present for a delayed run — the earliest time it could be claimed.
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        completed_at: { type: string, format: date-time }
        result:
          oneOf:
            - type: "null"
            - type: object
              properties:
                target_status: { type: integer }
                target_response:
                  description: Inline JSON result (results up to 256 KiB).
                href:
                  type: string
                  description: Download path of a file result.
                content_type: { type: string }
        error:
          oneOf:
            - type: "null"
            - type: object
              required: [code, message]
              properties:
                code:
                  type: string
                  enum: [invalid_payload, unknown_target, unknown_resolvent, resource_error,
                         invalid_resource_response, target_error, target_job_failed, target_timeout,
                         max_attempts_exceeded, internal]
                message: { type: string }
    ScheduleRequest:
      type: object
      required: [target, payload, cron]
      properties:
        target: { type: string }
        payload:
          type: object
          description: The body every run resolves and forwards.
        cron:
          type: string
          description: Five fields (`30 6 * * *`) or a descriptor (`@daily`, `@every 6h`).
        timezone:
          type: string
          default: UTC
          description: IANA zone name the expression is evaluated in.
        priority:
          type: integer
          minimum: -10
          maximum: 10
          default: 0
        options:
          type: object
          properties:
            cache:
              type: string
              enum: [use, bypass, refresh]
              description: Defaults to `refresh` for runs, so every run picks up fresh inputs.
        api_key:
          type: string
          deprecated: true
          description: Fallback for the `X-API-Key` header.
    Schedule:
      type: object
      required: [id, target, cron, timezone, next_run_at, created_at, updated_at, links]
      properties:
        id: { type: string }
        target: { type: string }
        cron: { type: string }
        timezone: { type: string }
        priority:
          type: integer
          description: Present when not the default 0.
        options:
          type: object
          properties:
            cache: { type: string, enum: [bypass, refresh] }
        next_run_at: { type: string, format: date-time }
        last_run_at: { type: string, format: date-time }
        last_job_id:
          type: string
          description: The request created for the most recent due time.
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        links:
          type: object
          properties:
            self: { type: string }
            runs: { type: string }
...
    JobList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/Job" }
        next_cursor:
          type: string
          description: Present only when more items match.
...